Move example code to examples
