--- examples/package.json
@@ -1,4 +1,0 @@
-{
-  "name": "quinn.examples",
-  "main": "./app.js"
-}
